services/collectors/requirements.txt
redis[hiredis]>=5.0.0
psutil>=5.9.0
pyroute2>=0.7.0
pydantic>=2.0.0
structlog>=23.0.0
pyyaml>=6.0.0
